LayoutTests/editing/selection/ios/update-selection-after-iframe-scroll-expected.txt

PASS selectionRectsBefore[0].top is 74
PASS selectionRectsBefore[0].width is 184
PASS selectionRectsBefore[0].left is 10
PASS selectionRectsBefore[0].height is 19
PASS selectionRectsBefore[1].top is 93
PASS selectionRectsBefore[1].width is 167
PASS selectionRectsBefore[1].left is 10
PASS selectionRectsBefore[1].height is 20
PASS selectionRectsAfter[0].top is 24
PASS selectionRectsAfter[0].width is 184
PASS selectionRectsAfter[0].left is 10
PASS selectionRectsAfter[0].height is 19
PASS selectionRectsAfter[1].top is 43
PASS selectionRectsAfter[1].width is 167
PASS selectionRectsAfter[1].left is 10
PASS selectionRectsAfter[1].height is 20
PASS successfullyParsed is true

TEST COMPLETE

This test verifies that a selection in an iframe area is kept up to date after scrolling. To test manually, tap the frame and scroll the editable area down; the selection move to account for the new scroll position.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
